refactoro: gnb 구조 변경 및 라우팅 대응
