+++ components/gnb/gnb.tsx
@@ -1,54 +1,20 @@
 'use client'
 
 import classNames from 'classnames/bind'
-import { deleteCookie } from 'cookies-next'
 
 import Image from 'next/image'
 import Link from 'next/link'
-import { usePathname, useRouter } from 'next/navigation'
 
 import useBreakpoint from '@/hooks/useBreakpoint.hooks'
 import useScrolledState from '@/hooks/useScrolledState'
 
-import { logout } from '@/api/axios/requestHandler/auth/post.apis'
-
-import { UserType } from '@/types/user.enum'
-
-import { btnTextHelpers } from './gnb.helpers'
 import styles from './gnb.module.scss'
-import GnbChip from './gnbChip'
-import { GnbChipStyle } from './gnbChip.types'
 
 const cx = classNames.bind(styles)
 
-export default function Gnb() {
-  const router = useRouter()
-  const pathname = usePathname()
+export default function Gnb({ children }: { children: React.ReactNode }) {
   const isTablet = useBreakpoint({ query: '(max-width: 1199px)' })
   const scrolled = useScrolledState()
-
-  const currentBtnText = btnTextHelpers(pathname)
-
-  const user = {
-    userType: pathname.includes('jff') ? 'forFun' : 'forDev',
-  }
-
-  const handleSignOut = async () => {
-    try {
-      const res = await logout()
-
-      if (res === null) {
-        deleteCookie('act')
-        deleteCookie('rft')
-        deleteCookie('uid')
-        deleteCookie('utp')
-
-        router.push('/')
-      }
-    } catch (err) {
-      throw new Error()
-    }
-  }
 
   return (
     <nav className={cx('gnbWrapper', { hasBorder: !isTablet }, { scrolled })}>
@@ -65,29 +31,7 @@ export default function Gnb() {
             </h1>
           </Link>
         </div>
-        <div className={styles.right}>
-          {currentBtnText === '로그인' ? (
-            <Link href="/signin" className={styles.login}>
-              <h2>로그인</h2>
-            </Link>
-          ) : (
-            <>
-              {user.userType === UserType.FORFUN && (
-                <span className={styles.gnbChip}>
-                  <GnbChip style={GnbChipStyle.LIGHTBLUE}>For Fun</GnbChip>
-                </span>
-              )}
-              {user.userType === UserType.FORDEV && (
-                <span className={styles.gnbChip}>
-                  <GnbChip style={GnbChipStyle.DEEPBLUE}>For Dev</GnbChip>
-                </span>
-              )}
-              <button className={styles.logout} onClick={handleSignOut}>
-                <h2>로그아웃</h2>
-              </button>
-            </>
-          )}
-        </div>
+        {children}
       </div>
     </nav>
   )
